Minimize api calling, fix ViewCount

+++ DataBaseAPI/POCO/VideoItemPOCO.cs
@@ -14,7 +14,7 @@ namespace DataBaseAPI.POCO
 
         public string Description { get; set; }
 
-        public int ViewCount { get; set; }
+        public long ViewCount { get; set; }
 
         public int Duration { get; set; }
