+++ Interfaces/Models/IVideoItem.cs
@@ -13,7 +13,7 @@ namespace Interfaces.Models
 
         string Description { get; set; }
 
-        int ViewCount { get; set; }
+        long ViewCount { get; set; }
 
         int Duration { get; set; }
