+++ SitesAPI/POCO/VideoItemPOCO.cs
@@ -19,7 +19,7 @@ namespace SitesAPI.POCO
 
         public string Description { get; set; }
         
-        public int ViewCount { get; set; }
+        public long ViewCount { get; set; }
 
         public int Duration { get; set; }
